+++ panel/scene-view/scene-view.html
@@ -30,7 +30,7 @@
     </template>
 
     <script type="text/javascript">
-        Editor.require('packages://scene/panel/scene-view/page-init');
+        Editor.require('packages://scene/panel/scene-view/init');
     </script>
     <script type="text/javascript" src="scene-view.js"></script>
 </dom-module>
